restructure api
closes #2
