+++ lib/webgl.js
@@ -222,32 +222,3 @@ WebGL.prototype.bindDestinationTexture = function(M, N, dstTex) {
 
 	return this.framebuffer;
 };
-
-/* create a typed array from a 2D javascript array */
-WebGL.fromArray = function(array, type, tranpose) {
-	var shape = [],
-			data,
-			c;   // number of columns
-
-	if(!tranpose){
-		shape[0] = array.length;
-		shape[1] = array[0].length;
-	} else {
-		shape[1] = array.length;
-		shape[0] = array[0].length;
-	}
-	c = shape[1];
-
-	type = type || Float32Array;
-
-	data = new type(shape[0]*shape[1]);
-
-	for (var ii = 0; ii < shape[0]; ++ii)
-		for (var jj = 0; jj < shape[1]; ++jj)
-		if(!tranpose)
-			data[ii*c + jj] = array[ii][jj];
-		else
-			data[ii*c + jj] = array[jj][ii];
-
-	return data;
-};
